+++ eslint.config.js
@@ -10,9 +10,10 @@ import pluginSecurityNode from 'eslint-plugin-security-node';
 // TODO wait for eslint-plugin-import to be usable in eslint v9; corresponding
 // issue see https://github.com/import-js/eslint-plugin-import/issues/2948
 
-export default [{
-  ignores: ["**/.vscode/"],
-  }, 
+export default [
+  {
+    ignores: ['**/.vscode/'],
+  },
   pluginJs.configs.recommended,
   pluginJsdoc.configs['flat/recommended'],
   pluginMocha.configs.flat.recommended,
@@ -28,7 +29,7 @@ export default [{
       },
     },
     plugins: {
-      'security-node': pluginSecurityNode
+      'security-node': pluginSecurityNode,
     },
     rules: {
       'mocha/no-mocha-arrows': 'off',
@@ -37,7 +38,7 @@ export default [{
       ...pluginSecurityNode.configs.recommended.rules,
       // disable redundant checks (already exists as security/detect-possible-timing-attacks)
       'security-node/detect-possible-timing-attacks': 'off',
-      'security-node/non-literal-reg-expr': 'off'
+      'security-node/non-literal-reg-expr': 'off',
     },
   },
   {