+++ scripts/readme-updater.cjs
@@ -1,8 +1,9 @@
 // readme-updater.js
 // standard-version updater for readme files with iconified semantic version string.
 
-module.exports.readVersion = function(contents) {
-  const versionSearchPattern = /(.*!\[Version]\(https:\/\/img.shields.io\/badge\/version-)(.+)(-blue\.svg.*)/is;
+module.exports.readVersion = function (contents) {
+  const versionSearchPattern =
+    /(.*!\[Version]\(https:\/\/img.shields.io\/badge\/version-)(.+)(-blue\.svg.*)/is;
   const regexResult = contents.match(versionSearchPattern);
   let version = '';
   if (regexResult.length === 4) {
@@ -12,8 +13,9 @@ module.exports.readVersion = function(contents) {
   return version;
 };
 
-module.exports.writeVersion = function(contents, version) {
-  const versionSearchPattern = /(.*!\[Version]\(https:\/\/img.shields.io\/badge\/version-)(.+)(-blue\.svg.*)/is;
+module.exports.writeVersion = function (contents, version) {
+  const versionSearchPattern =
+    /(.*!\[Version]\(https:\/\/img.shields.io\/badge\/version-)(.+)(-blue\.svg.*)/is;
   const regexResult = contents.match(versionSearchPattern);
   let newContents = contents;
   if (regexResult.length === 4) {